tests/requirements.txt
# Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-benchmark==4.0.0
pytest-timeout==2.2.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
scipy==1.11.4
numpy==1.26.2

# STT/TTS
openai-whisper==20231117
jiwer==3.0.3
gtts==2.5.0

# Quality Metrics
pesq==0.0.4
pystoi==0.3.3

# Optimization
scikit-optimize==0.9.0
optuna==3.5.0
bayesian-optimization==1.4.3

# Monitoring & Metrics
influxdb==5.3.1
prometheus-client==0.19.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
websockets==12.0
aiohttp==3.9.1
